feat: V6.0 mechanism + V5.0 stats + 5 code-review fixes
BREAKTHROUGH: V5.0 dichotomy now has a mechanism-backed theorem.

V6.0 — depth sweep confirms the Hanin & Nica 2020 log-normal prediction
(Comm. Math. Phys. 376, 287-322, arXiv:1812.05994). 7 depths x 5 seeds on
untrained ReLU MLPs:
  H1: Var[log F_ii] linear in L   -> slope 5.73, R^2 = 0.906 (PASS)
  H2: log(T1/T3) linear in sqrt(L) -> slope 11.5, R^2 = 0.983 (PASS)
At L=20, observed T1/T3 = 9e15 (theoretical log-normal-tail prediction).

V5.0-stats — bootstrap 95% CIs + Mann-Whitney U test on the 10-system
dichotomy: p = 5.1e-17, rank-biserial r = 1.000 (COMPLETE rank separation,
n_deep=46, n_rest=47). No CI in the deep-sequential group crosses below
100; no CI in the rest group crosses above 100.

Code-review findings fixed:
1. lattice_u1 grad_action had wrong sign on backward plaquettes in BOTH
   branches; corrected to beta * (sin_fwd - sin_bwd) with full derivation
   in docstring. Re-running lattice_u1 (seed 0: T1/T3=1.6, was 2.0 --
   still in O(1) band, dichotomy unaffected).
2. learning_baselines gaussian_process_fim was computing 1/lambda
   (posterior precision), not Fisher information. Replaced with the
   correct FIM diagonal formula K^2/sigma^2 (same as kernel_ridge).
   GP T1/T3: 5.37 -> 1.97 (still in O(1-5) band).
3. analyze.py FIM accumulator promoted from float32 to float64 to
   prevent tier-3 underflow (min normal ~1.2e-38).
4. experiment_pytorch.py same float32 -> float64 promotion.
5. scaling_experiment_extended.py same promotion.
6. run_trained_vs_untrained.py --probes default 200 -> 2000.
7. baselines.py Ising + CA docstrings clarified re: squaring convention.
8. lattice_u1.py thermalise_s JSON field now records actual duration,
   not a timestamp. --decorr CLI flag added (default 20, was hardcoded 5).

Paper/README corrections:
- FIM CV at W=14000 from stale 5-seed value (1.85%) to canonical 20-seed
  value (1.51%) in paper, README, findings.
- GP T1/T3 5.37 -> 1.97 in paper table, README table.
- Abstract exponent band '1.0-1.5' -> '1.4-2.8 (300p) rising to 2.3-5.5
  (2000p)' matching the actual canonical values in the 4-task universality
  table.
- findings.md retraction note on stale T4 1.067 'clean 4-width' exponent.
- Paper S2 Related Work expanded with Karakida-Amari (pathological FIM
  spectrum), Papyan (3-level Hessian hierarchy), Hanin-Nica (PoRM
  log-normal), Yang-Hu (NTK/muP dichotomy), Frankle-Carbin (lottery
  tickets), Swingle/Susskind-Brown (tensor-network / circuit-complexity
  holographic precursors).

New files:
- docs/v6_0_mechanism_hanin_nica.md — full derivation + V6.0 empirical
  confirmation.
- experiments/v6_0_depth_mechanism/ — depth sweep script + JSON.
- experiments/v5_0_dichotomy_stats/ — bootstrap + Mann-Whitney script,
  JSON, markdown summary.

+++ experiment_pytorch.py
@@ -83,11 +83,16 @@ for name, param in net.named_parameters():
         })
         print(f"  {name}: top3={[round(s,3) for s in top3]}, ratio={ratio:.1f}, std={float(S.std()):.4f}")
 
-# --- Measurement 2: FIM Eigenvalue Spectrum ---
+# --- Measurement 2: FIM Diagonal ---
 print("\n=== Measurement 2: FIM Diagonal Approximation ===")
-# Compute empirical FIM diagonal: F_ii = E[(d log p / d theta_i)^2]
+# Empirical FIM diagonal: F_ii = E[(d log p / d theta_i)^2].
+# Accumulate in float64: grad^2 for tier-3 params can fall below the float32
+# minimum normal (~1.2e-38) and round to zero, which inflates the reported
+# T1/T3 ratio. See experiments/v2_1_qec/analyze.py for the canonical version.
 net.eval()
-fim_diag = {name: torch.zeros_like(p) for name, p in net.named_parameters()}
+fim_diag = {
+    name: torch.zeros_like(p, dtype=torch.float64) for name, p in net.named_parameters()
+}
 
 n_samples = 1000
 for _ in range(n_samples):
@@ -98,7 +103,7 @@ for _ in range(n_samples):
     loss.backward()
     for name, p in net.named_parameters():
         if p.grad is not None:
-            fim_diag[name] += p.grad.data ** 2
+            fim_diag[name] += p.grad.data.double() ** 2
 
 # Normalize
 for name in fim_diag:
